# docs/swagger.yaml
basePath: /api
definitions:
  models.Answer:
    properties:
      created_at:
        type: string
      id:
        type: integer
      question_id:
        type: integer
      text:
        type: string
      user_id:
        type: string
    type: object
  models.Question:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.Answer'
        type: array
      created_at:
        type: string
      id:
        type: integer
      text:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Service for managing questions and answers
  title: Questions and Answers API
  version: "1.0"
paths:
  /api/answers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific answer by its ID
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an answer
      tags:
      - answers
    get:
      consumes:
      - application/json
      description: Get a specific answer by its ID
      parameters:
      - description: Answer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Answer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get answer by ID
      tags:
      - answers
  /api/questions:
    get:
      consumes:
      - application/json
      description: Get a list of all questions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Question'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all questions
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Create a new question
      parameters:
      - description: Question object
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/models.Question'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Question'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new question
      tags:
      - questions
  /api/questions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a question
      tags:
      - questions
    get:
      consumes:
      - application/json
      description: Get a specific question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Question'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get question by ID
      tags:
      - questions
  /api/questions/{question_id}/answers:
    post:
      consumes:
      - application/json
      description: Create a new answer for a specific question
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: integer
      - description: Answer object
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/models.Answer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Answer'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new answer for a question
      tags:
      - answers
swagger: "2.0"